Cosmetic change on lru_cache backport
